Only intercept DNS traffic
DNS66 only expects DNS traffic to go through the vpn, so any other traffic just gets dropped, which is why we can't intercept all traffic like with Virtual-Hosts. Instead, we only addRoute() for every existing DNS server so only that traffic is intercepted, and that traffic is sucessfully forwarded.

Galaxy phones also seem to handle Vpn services differently - this seems to be the solution.