Use most up-to-date hosts files downloaded when starting the filter.
Previously, we would use whatever hosts files were downloaded. If there was no internet connectivity and all the files failed to download, we would only use the built in one, but when reconnecting to the internet, we would still only be using the built in one and not any of the ones the user selected. Now, we store the downloaded file as the hashcode of the url, and load whichever url/file is enabled.